pyproject.toml
[project]
name = "quartz-council"
version = "0.1.0"
description = "Opt-in AI pull request reviewer with multi-agent review council"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "openai>=1.50.0",
    "pydantic>=2.9.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quartzcouncil"]
